Refactor solar cards and UI components for cleaner architecture
Core changes:
- Remove solar_cards.py module (consolidate into solar_viz)
- Remove solar_ui_service.py (move business logic to core modules)
- Update test assertions for Wh units in 15min heatmap (kWh * 1000)
- Adjust heatmap figure heights from 1000px to 500px for better proportions
- Remove unused get_period_stats function and clear selection test button

Rationale: Consolidates scattered UI components into core modules, eliminates redundant service layer, and fixes unit display consistency in heatmaps.
